requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
confluent-kafka>=2.4
influxdb-client>=1.44
psycopg2-binary>=2.9
langchain-core>=0.2
langchain-google-genai>=1.0
langgraph>=0.2
openai>=1.0
streamlit>=1.31
python-dotenv>=1.0
paho-mqtt>=1.6

websockets~=15.0.1
altair~=6.0.0
pandas~=2.3.3
